osr2png/requirements.txt
aiohttp
Pillow
PySimpleGUI
requests
autologging
logger